refactor: 루트 페이지 구현으로 Hydration 에러 방지 목적 레이아웃 수정
