+++ frontend/src/app/admin/layout.tsx
@@ -1,16 +1,6 @@
-import { Geist, Geist_Mono } from "next/font/google";
-import "../bookbook/globals.css";
+import "../globals.css";
 import ContextLayout from "./contextLayout";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
+import { ToastContainer } from "react-toastify";
 
 export default function RootLayout({
   children,
@@ -18,10 +8,12 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en">
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased flex flex-col min-h-screen`}>
-        <ContextLayout>{children}</ContextLayout>
-      </body>
-    </html>
+    <>
+      <ToastContainer
+        position="top-right"
+        autoClose={3000}
+      />
+      <ContextLayout>{children}</ContextLayout>
+    </>
   );
 }
