+++ frontend/src/app/bookbook/layout.tsx
@@ -1,39 +1,21 @@
-import type { Metadata } from 'next';
-import { Geist, Geist_Mono } from 'next/font/google';
-import './globals.css';
+import '../globals.css';
 import Bottom from '../components/Bottom';
 import ClientLayout from './ClientLayout';
-import {ToastContainer} from "react-toastify";
+import { ToastContainer } from "react-toastify";
 
-const geistSans = Geist({
-  variable: '--font-geist-sans',
-  subsets: ['latin'],
-});
-const geistMono = Geist_Mono({
-  variable: '--font-geist-mono',
-  subsets: ['latin'],
-});
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-};
 export default function RootLayout({
-                                     children,
-                                   }: Readonly<{
+  children,
+}: Readonly<{
   children: React.ReactNode;
 }>) {
   return (
-      <html lang="en">
-      <body
-          className={`${geistSans.variable} ${geistMono.variable} antialiased flex flex-col min-h-screen`}
-      >
+    <>
       <ClientLayout>{children}</ClientLayout>
       <Bottom />
       <ToastContainer
           position="bottom-center"
           autoClose={3000}
       />
-      </body>
-      </html>
+    </>
   );
 }
